requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
